requirements.txt
autopep8==1.6.0
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
Flask==2.0.2
Flask-Cors==3.0.10
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
psycopg2-binary==2.9.3
pycodestyle==2.8.0
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
toml==0.10.2
urllib3==1.26.8
Werkzeug==2.0.2
